calcagent/requirements.txt
google-cloud-aiplatform[adk,agent_engines]
cloudpickle==3.1.1
pydantic==2.11.7
mcp
google-adk
